IAM roles support for the S3 Client
If the passed in authenticationInfo is null, use the AWS SDK provided
provider chain that covers IAM roles, ECS container roles, etc. instead
of passing the null as credentials. Passing null as credentials
shortcuts the provider chain and puts the client in an anonymous mode.
Instead, we now always go through the provider chain which, if it can't
find any credentials in the chain, throws an exception (instead of
anonymous mode).